Created and configured new example module for new API
This new example module is configured for the new API
It will display a message on any exposed navbars or footers on themes

This is just an example and is not designed to be functional.
However, this does demonstrate the new API for other modules.

This is not final and major (even breaking) changes could be made soon