Stopped being so dramatic over failed CSRF
The CSRF failed message is a bit less over the top and makes it more clear what happened
We will now give advice to the user that if they were trying to perform an action, just try again.
However, we made it clear still that if they clicked  a link, they should close the page.