Repurpose the testing themes/modules
The built in theme "testing_theme" has been included in DSJAS since the first commit
To be honest, it is only here because I forgot to tell git to ignore it.
But now, I've decided to make use of it.

The test theme is now deleted by the package script
This essentially prevents it from being shipped to real program users

However, it is still in source control
So, it now just contains a nice greeting for developers on the dev version

As a related change, the example module will now only ship to developers
It is deleted by the same script as the testing theme