Add visual loading spinner to final installer page
The final stage of the installer starts a long-running operation
without any visual indication of this to the user at all.
This led some users to think that nothing was happening - as in
issue #2

Added an animated spinner to the submit button which shows to the
user that something is happening and that the button is actually
working.
Added JavaScript logic which changes the spinner state based on the
current state of the request.