Developer containers: add new developer PHP.ini
Added a separate PHP.ini file for settings specific to devcontainer
setups.

The first difference, in this version at least, is that the
containers will now contain the capability to log errors to your
workspace root (out of the web tree). This will be really convenient
for me and others working out of containers.

At some point in the future, the configs will diverge further. For
now, they are almost identical, aside from one or two values.