Fix misspelled table name in database setup script
The installer was using the name 'siteUsers' but the rest of DSJAS
was using 'siteusers'. The installer will now use the correct
spelling/capitalization

This bug wasn't caught for a long time, due to the fact that
different database servers handle the table name capitalization
differently. My database server, for instance, was treating them
all as lowercase rather than respecting the given values

For the sake of concistency and to prevent issues like #2 from ever
happening again, table names will only ever be lower case from this
point forward