Add definition option to opt out of bootstrap load
Scripts can now opt out of the loading of Bootstrap.css, Bootstrap.js
and all other CSS/JS libraries in the bootstrapper script by defining
"NOLOAD_BOOTSTRAP_HEAD" before including the bootstrapper script

This constant is checked for if it is defined. If not, the load
continues as normal. If it is, the load ignores the section of the
bootstrapper which adds the head element (and its contents)

Will be useful for AJAX scripts and API endpoints which still require
access to bootstrapper constants without the CSS and JS