Improvements to statistics summary UI
Added dynamic clock updating in JavaScript. The amount of time elapsed
in the session will now count up, emulating a clock
JavaScript will also now pad the output with zeros when a segment of the
clock is in single digits; i.e, "01:25:09" not "1:25:9"

Added dynamic updating of the page hits count using AJAX
A new remote API has been added under the admin directory which returns
a JSON object containing a count of the number of page hits
This API is only open to authenticated users

Finally, changed some of the default values of the counters when there
is no active session. Now, when it makes sense to do so, the counters
will display a simple zero rather than "N/A". The start time still
displays "N/A"
This changes makes the counters look more as though they have been
started when the session starts, rather than them having been created