Remove database cleanup from statistics destructor
The statistics manager was cleaning up the database connection manually
in its destructor.
This was causing a lot of errors, as PHP was cleaning up the database
connection before we were.

That problem leads to errors like "Failed to fetch MySQLi connection at
localhost", which can get messy and confuse people looking through error
logs