Add initial vanity account number system
Added a system to the database to store a "vanity account number": a
formatted and nicely generated account number which doesn't look out of
place on a banking dasboard.

They are not guaranteed to be unique (the use is purely visual) and they
are handled completely implicitly by DSJAS: the admin has no need to
intervene or set anything up manually.