Add new option to simulate absence of admin panel
Added a new stealth/security option to redirect all admin pages to a 404
page. This applies to all pages except admin login pages (allowing the
setting to still be disabled, if required).

This is a somewhat softer option than completely disabling the admin
panel using the "access to admin" option, meaning that many users who do
not want to edit their configuration on disk will want to opt for this
instead.