Redo admin dashboard stealth features
Remade the configuration file to remove the requirement for a flag to be
present for the admin dashboard to work. This should make the site more
resilient, as a malformed/missing config will no longer brick the admin
dashoard.

Re-wrote and re-formatted some UIs and logic to account for the change
and updated default config files.