Add theme API for interacting with statistics
Add an API to "theme/General.php", which allows themes to add and modify
statistics to their liking.
Themes can directly modify the values of number statistics, but may only
request DSJAS to update the values of other types, thereby avoiding
invalid statistics.
Themes have all the business of registering and categorising statistics
abstracted away: the only thing they need to know is a human-friendly
name, the requested type and value, and the (optional) category.