Begin to add proper database handling
Our system for dealing with prepared statements will be that each
package in the data system is responsible for self-compiling at database
startup. If this fails, the application panics (as this is developer
error). It isn't the cleanest, but it is thread safe and achieves what
we want it to.